#13 MOVE ScraperBloomberg to separate file
